feat: add service layer, ItemPedido entity and H2 configuration
  - Add CRUD services for Cliente, Endereco, Entregador, Pedido and Produto
  - Add ItemPedido entity with ManyToOne to Pedido and Produto
  - Add EnderecoRepository and ItemPedidoRepository
  - Fix repository generic types from Long to Integer (Entregador, Produto)
  - Normalize entity ID field names to id across all entities
  - Annotate all controllers with @RestController and @RequestMapping
  - Configure H2 in-memory database, JPA settings and server port 8081
  - Add devtools and H2 dependencies to pom.xml

+++ src/main/java/br/com/centrallanches/lanchonete_api/repository/EntregadorRepository.java
@@ -3,5 +3,5 @@ package br.com.centrallanches.lanchonete_api.repository;
 import br.com.centrallanches.lanchonete_api.entity.Entregador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntregadorRepository extends JpaRepository <Entregador, Long> {
+public interface EntregadorRepository extends JpaRepository <Entregador, Integer > {
 }
